Refactor to new folder structure
