+++ components/org.wso2.identity.webhook.common.event.handler/src/main/java/org/wso2/identity/webhook/common/event/handler/api/constant/Constants.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.identity.webhook.common.event.handler.constant;
+package org.wso2.identity.webhook.common.event.handler.api.constant;
 
 /**
  * Constants class.
