+++ components/org.wso2.identity.webhook.common.event.handler/src/main/java/org/wso2/identity/webhook/common/event/handler/internal/util/EventHookHandlerInternalUtils.java
@@ -1,22 +1,4 @@
-/*
- * Copyright (c) 2024, WSO2 LLC. (http://www.wso2.com).
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.identity.webhook.common.event.handler.util;
+package org.wso2.identity.webhook.common.event.handler.internal.util;
 
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
@@ -34,33 +16,30 @@ import org.wso2.carbon.identity.event.event.Event;
 import org.wso2.identity.event.common.publisher.model.EventContext;
 import org.wso2.identity.event.common.publisher.model.EventPayload;
 import org.wso2.identity.event.common.publisher.model.SecurityEventTokenPayload;
-import org.wso2.identity.webhook.common.event.handler.constant.Constants;
-import org.wso2.identity.webhook.common.event.handler.internal.EventHookHandlerDataHolder;
-import org.wso2.identity.webhook.common.event.handler.model.EventData;
+import org.wso2.identity.webhook.common.event.handler.api.constant.Constants;
+import org.wso2.identity.webhook.common.event.handler.api.model.EventData;
+import org.wso2.identity.webhook.common.event.handler.internal.service.EventHookHandlerDataHolder;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
-import javax.servlet.http.HttpServletRequest;
 
 import static org.wso2.carbon.identity.application.authentication.framework.util.FrameworkUtils.CORRELATION_ID_MDC;
 
-/**
- * This class contains the utility method implementations.
- */
-public class EventHookHandlerUtils {
+public class EventHookHandlerInternalUtils {
 
-    private static final Log log = LogFactory.getLog(EventHookHandlerUtils.class);
-    private static volatile EventHookHandlerUtils instance;
+    private static final Log log = LogFactory.getLog(EventHookHandlerInternalUtils.class);
+    private static volatile EventHookHandlerInternalUtils instance;
 
-    private EventHookHandlerUtils() {}
+    private EventHookHandlerInternalUtils() {}
 
-    public static EventHookHandlerUtils getInstance() {
+    public static EventHookHandlerInternalUtils getInstance() {
 
         if (instance == null) {
-            synchronized (EventHookHandlerUtils.class) {
+            synchronized (EventHookHandlerInternalUtils.class) {
                 if (instance == null) {
-                    instance = new EventHookHandlerUtils();
+                    instance = new EventHookHandlerInternalUtils();
                 }
             }
         }
@@ -109,6 +88,7 @@ public class EventHookHandlerUtils {
                 .build();
     }
 
+
     /**
      * Retrieve the audience.
      *
@@ -154,6 +134,10 @@ public class EventHookHandlerUtils {
         return correlationID;
     }
 
+
+
+
+
     private void setLocalUserClaimsToAuthenticatedUser(AuthenticatedUser authenticatedUser,
                                                        AuthenticationContext context) {
 
@@ -198,17 +182,6 @@ public class EventHookHandlerUtils {
         return null;
     }
 
-    /**
-     * Get the tenant qualified URL with path.
-     *
-     * @param endpoint Endpoint.
-     * @return Tenant qualified URL.
-     */
-    public String constructFullURLWithEndpoint(String endpoint) {
-
-        endpoint = constructBaseURL() + endpoint;
-        return endpoint;
-    }
 
     /**
      * Publish the event payload.
@@ -233,4 +206,5 @@ public class EventHookHandlerUtils {
                     eventUri), e);
         }
     }
+
 }
